refactor: CouponList의 useCoupon 로직과 UI로직 추상화
